+++ apps/ledger-live-mobile/src/screens/OperationDetails/Content.tsx
@@ -408,7 +408,11 @@ export default function Content({
 
       {uniqueSenders.length > 0 && (
         <View style={sectionStyles.wrapper}>
-          <DataList data={uniqueSenders} title={<Trans i18nKey="operationDetails.from" />} />
+          <DataList
+            testID="operationDetails-sender"
+            data={uniqueSenders}
+            title={<Trans i18nKey="operationDetails.from" />}
+          />
         </View>
       )}