+++ libs/coin-modules/coin-solana/src/transaction.ts
@@ -20,11 +20,10 @@ import {
   toTransactionStatusRawCommon as toTransactionStatusRaw,
 } from "@ledgerhq/coin-framework/serialization";
 import type { Account } from "@ledgerhq/types-live";
-import { getTokenById } from "@ledgerhq/cryptoassets/index";
+import { findTokenByAddressInCurrency } from "@ledgerhq/cryptoassets/index";
 import { findSubAccountById, getAccountCurrency } from "@ledgerhq/coin-framework/account";
 import { formatCurrencyUnit } from "@ledgerhq/coin-framework/currencies";
 import { assertUnreachable } from "./utils";
-import { toTokenId } from "./logic";
 
 export const fromTransactionRaw = (tr: TransactionRaw): Transaction => {
   const common = fromTransactionCommonRaw(tr);
@@ -72,7 +71,7 @@ function formatCommand(mainAccount: Account, tx: Transaction, command: Command) 
     case "token.transfer":
       return formatTokenTransfer(mainAccount, tx, command);
     case "token.createATA":
-      return formatCreateATA(command);
+      return formatCreateATA(mainAccount, command);
     case "stake.createAccount":
       return formatStakeCreateAccount(mainAccount, tx, command);
     case "stake.delegate":
@@ -147,8 +146,11 @@ function formatTokenTransfer(mainAccount: Account, tx: Transaction, command: Tok
   return "\n" + str;
 }
 
-function formatCreateATA(command: TokenCreateATACommand) {
-  const token = getTokenById(toTokenId(command.mint));
+function formatCreateATA(mainAccount: Account, command: TokenCreateATACommand) {
+  const token = findTokenByAddressInCurrency(command.mint, mainAccount.currency.id);
+  if (!token) {
+    throw new Error(`token for mint "${command.mint}" not found`);
+  }
   const str = [`  OPT IN TOKEN: ${token.ticker}`].filter(Boolean).join("\n");
   return "\n" + str;
 }