+++ libs/ledgerjs/packages/types-live/src/feature.ts
@@ -205,6 +205,8 @@ export type Features = CurrencyFeatures & {
   llmRebornLP: Feature_LlmRebornLP;
   llmRebornFlex: DefaultFeature;
   llmAccountListUI: DefaultFeature;
+  llmLedgerSyncEntryPoints: Feature_LlmLedgerSyncEntryPoints;
+  lldLedgerSyncEntryPoints: Feature_LldLedgerSyncEntryPoints;
 };
 
 /**
@@ -353,7 +355,6 @@ export type Feature_ProtectServicesMobile = Feature<{
     };
   };
   account: {
-    loginURI: string;
     homeURI: string;
   };
   protectId: string;
@@ -383,7 +384,6 @@ export type Feature_ProtectServicesDesktop = Feature<{
   };
   account: {
     homeURI: string;
-    loginURI: string;
   };
   protectId: string;
 }>;
@@ -530,6 +530,18 @@ export type Feature_LlmWalletSync = Feature<{
   environment: WalletSyncEnvironment;
   watchConfig: WalletSyncWatchConfig;
   learnMoreLink: string;
+}>;
+
+export type Feature_LlmLedgerSyncEntryPoints = Feature<{
+  manager: boolean;
+  accounts: boolean;
+  settings: boolean;
+}>;
+export type Feature_LldLedgerSyncEntryPoints = Feature<{
+  manager: boolean;
+  accounts: boolean;
+  settings: boolean;
+  onboarding: boolean;
 }>;
 
 export type Feature_LlCounterValueGranularitiesRates = Feature<{